api/requirements.txt
azure-functions
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
python-dotenv
